requirements.txt
PyYAML==6.0.2
openai==0.27.0
requests==2.32.3
ollama==0.4.5
